Parent script to invoke subcommands

+++ install.sh
@@ -1,3 +1,4 @@
+[ -e /usr/local/vimlink ] && exit 0
 git clone git://github.com/mnoble/vimlink /usr/local/vimlink
 ln -s /usr/local/vimlink/bin/vimlink-add /usr/local/bin/vimlink-add
 ln -s /usr/local/vimlink/bin/vimlink-install /usr/local/bin/vimlink-install
